Update blanket write_wkt impl
